split legacy graph execution tests
