+++ docs/COMPLETION_AUDIT.md
@@ -311,6 +311,10 @@ and do not assume Phase 4 is ready without evidence.
   `tests/integration/cli_build/legacy_graph_command_host_effects/file_reads.rs`,
   leaving env-effect ordering and unrelated test skip coverage in the parent
   module.
+- Legacy `build-graph build.zen` execution tests now split single-target,
+  multi-target, and dependency-order compilation coverage into
+  `tests/integration/cli_build/legacy_graph_command/execution.rs`, leaving CLI
+  usage and legacy JSON diagnostics in the parent module.
 - Direct `zen build.zen` host-effect integration tests now split declared
   file-read fallback accept/reject coverage into
   `tests/integration/cli_build/direct_build_graph_host_effects/file_reads.rs`,
