+++ docs/PHASE_PLAN.md
@@ -488,6 +488,9 @@ checked-in docs, tests, and commits only.
 - Legacy `build-graph build.zen` host-effect tests now keep declared file-read
   fallback accept/reject cases in a focused child module, leaving env-effect
   ordering and unrelated test skip coverage in the parent module.
+- Legacy `build-graph build.zen` execution tests now keep single-target,
+  multi-target, and dependency-order compilation coverage in a focused child
+  module, leaving CLI usage and legacy JSON diagnostics in the parent module.
 - Direct `zen build.zen` host-effect tests now keep declared file-read fallback
   accept/reject cases in a focused child module, leaving env-effect ordering and
   unrelated test skip coverage in the parent integration module.
